+++ instrumentation/opentelemetry-instrumentation-django/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Django instrumentation is now enabled by default but can be disabled by setting `OTEL_PYTHON_DJANGO_INSTRUMENT` to `False` ([#1239](https://github.com/open-telemetry/opentelemetry-python/pull/1239))
+
 ## Version 0.14b0
 
 Released 2020-10-13